Add container to ApprovableTypeTab to avoid passing icon as a property
