+++ src/pages/Admin/Home/Dashboard/Menu/ApprovableTypeTab/interfaces.ts
@@ -1,0 +1,20 @@
+import { ApprovableEntityType } from "$interfaces/Approvable";
+import { FunctionComponent } from "react";
+
+export interface IApprovableTypeTabContainerProps {
+  iconTitle: string;
+  types: ApprovableEntityType[];
+  type: ApprovableEntityType;
+  onClick: (selected: boolean, entityType: ApprovableEntityType) => void;
+}
+
+export interface IApprovableTypeTabProps {
+  iconTitle: string;
+  Icon: FunctionComponent<IIcon>;
+  toggleSelected: () => void;
+  selected: boolean;
+}
+
+interface IIcon {
+  className?: string;
+}
