+++ src/pages/Admin/Home/Dashboard/Menu/TypeFilter/component.tsx
@@ -1,7 +1,5 @@
 import React, { FunctionComponent } from "react";
 import { ApprovableTypeTab } from "../ApprovableTypeTab";
-import { CompanyIcon } from "../../CompanyIcon";
-import { ApplicantIcon } from "../../ApplicantIcon";
 import { ITypeFilterProps } from "./interfaces";
 import { APPLICANT, COMPANY } from "$typenames";
 export const TypeFilter: FunctionComponent<ITypeFilterProps> = (
@@ -13,14 +11,12 @@ export const TypeFilter: FunctionComponent<ITypeFilterProps> = (
 ) => (
   <>
     <ApprovableTypeTab
-      Icon={CompanyIcon}
       iconTitle={translations.companyIconTitle}
       types={types}
       type={COMPANY}
       onClick={toggleType}
     />
     <ApprovableTypeTab
-      Icon={ApplicantIcon}
       iconTitle={translations.companyIconTitle}
       types={types}
       type={APPLICANT}
